services/soar-orchestrator/requirements.txt
kafka-python==2.0.2
opensearch-py==2.3.1
pyyaml==6.0.1
requests==2.31.0
celery==5.3.1
redis==4.6.0
jinja2==3.1.2
psycopg2-binary==2.9.7